Fixed created_by and updated_by not being saved.
As instance.created_by and instance.updated_by in managers.py was set
after the .create() call, the fields were never actually set in the
database. They were, however, returned.
The previous tests succeded, because they only checked the data they got
back, and not what was in the database.